Added serialization tests for complex types. Fix test data where decimal was actually a double.

+++ tests/MediumTests/SerializationAndConversionTests/StructVos/Types/DecimalVo.cs
@@ -1,29 +1,29 @@
 ﻿namespace Vogen.IntegrationTests.TestTypes.StructVos
 {
-    [ValueObject(conversions: Conversions.None, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.None, underlyingType: typeof(decimal))]
     public partial struct DecimalVo { }
 
-    [ValueObject(conversions: Conversions.None, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.None, underlyingType: typeof(decimal))]
     public partial struct NoConverterDecimalVo { }
 
-    [ValueObject(conversions: Conversions.TypeConverter, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.TypeConverter, underlyingType: typeof(decimal))]
     public partial struct NoJsonDecimalVo { }
 
-    [ValueObject(conversions: Conversions.NewtonsoftJson, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.NewtonsoftJson, underlyingType: typeof(decimal))]
     public partial struct NewtonsoftJsonDecimalVo { }
 
-    [ValueObject(conversions: Conversions.SystemTextJson, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.SystemTextJson, underlyingType: typeof(decimal))]
     public partial struct SystemTextJsonDecimalVo { }
 
-    [ValueObject(conversions: Conversions.NewtonsoftJson | Conversions.SystemTextJson, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.NewtonsoftJson | Conversions.SystemTextJson, underlyingType: typeof(decimal))]
     public partial struct BothJsonDecimalVo { }
 
-    [ValueObject(conversions: Conversions.EfCoreValueConverter, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.EfCoreValueConverter, underlyingType: typeof(decimal))]
     public partial struct EfCoreDecimalVo { }
 
-    [ValueObject(conversions: Conversions.DapperTypeHandler, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.DapperTypeHandler, underlyingType: typeof(decimal))]
     public partial struct DapperDecimalVo { }
 
-    [ValueObject(conversions: Conversions.LinqToDbValueConverter, underlyingType: typeof(double))]
+    [ValueObject(conversions: Conversions.LinqToDbValueConverter, underlyingType: typeof(decimal))]
     public partial struct LinqToDbDecimalVo { }
 }
